Move event_listener module into evm::tracing
