+++ evm_loader/api/src/main.rs
@@ -11,7 +11,6 @@ pub use neon_lib::commands;
 pub use neon_lib::config;
 pub use neon_lib::context;
 pub use neon_lib::errors;
-pub use neon_lib::event_listener;
 pub use neon_lib::rpc;
 pub use neon_lib::syscall_stubs;
 pub use neon_lib::types;
