+++ evm_loader/lib/src/lib.rs
@@ -3,7 +3,6 @@ pub mod commands;
 pub mod config;
 pub mod context;
 pub mod errors;
-pub mod event_listener;
 pub mod rpc;
 pub mod syscall_stubs;
 pub mod types;
