+++ evm_loader/lib/src/types/mod.rs
@@ -11,99 +11,15 @@ use tokio::runtime::Runtime;
 use tokio::task::block_in_place;
 pub use tracer_ch_db::{ChError, ChResult, ClickHouseDb as TracerDb};
 
-use crate::event_listener::trace::{TraceCallConfig, TraceConfig};
+use evm_loader::evm::tracing::event_listener::trace::{TraceCallConfig, TraceConfig};
+use evm_loader::types::Bytes;
 use {
     ethnum::U256,
-    hex::FromHex,
     postgres::NoTls,
-    serde::{de::Visitor, Deserialize, Deserializer, Serialize, Serializer},
-    std::{fmt, ops::Deref},
+    serde::{Deserialize, Deserializer, Serialize, Serializer},
     thiserror::Error,
     tokio_postgres::{connect, Client},
 };
-
-/// Wrapper structure around vector of bytes.
-#[derive(Debug, PartialEq, Eq, Default, Hash, Clone)]
-pub struct Bytes(pub Vec<u8>);
-
-impl Bytes {
-    /// Simple constructor.
-    pub fn new(bytes: Vec<u8>) -> Bytes {
-        Bytes(bytes)
-    }
-}
-
-impl Deref for Bytes {
-    type Target = Vec<u8>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl From<Vec<u8>> for Bytes {
-    fn from(bytes: Vec<u8>) -> Bytes {
-        Bytes(bytes)
-    }
-}
-
-impl From<Bytes> for Vec<u8> {
-    fn from(value: Bytes) -> Self {
-        value.0
-    }
-}
-
-impl Serialize for Bytes {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let mut value = "0x".to_owned();
-        value.push_str(hex::encode(&self.0).as_str());
-        serializer.serialize_str(value.as_ref())
-    }
-}
-
-impl<'a> Deserialize<'a> for Bytes {
-    fn deserialize<D>(deserializer: D) -> Result<Bytes, D::Error>
-    where
-        D: Deserializer<'a>,
-    {
-        deserializer.deserialize_any(BytesVisitor)
-    }
-}
-
-struct BytesVisitor;
-
-impl<'a> Visitor<'a> for BytesVisitor {
-    type Value = Bytes;
-
-    fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        write!(formatter, "a 0x-prefixed, hex-encoded vector of bytes")
-    }
-
-    fn visit_str<E>(self, value: &str) -> Result<Self::Value, E>
-    where
-        E: serde::de::Error,
-    {
-        if value.len() >= 2 && value.starts_with("0x") && value.len() & 1 == 0 {
-            Ok(Bytes::new(FromHex::from_hex(&value[2..]).map_err(|e| {
-                serde::de::Error::custom(format!("Invalid hex: {e}"))
-            })?))
-        } else {
-            Err(serde::de::Error::custom(
-                "Invalid bytes format. Expected a 0x-prefixed hex string with even length",
-            ))
-        }
-    }
-
-    fn visit_string<E>(self, value: String) -> Result<Self::Value, E>
-    where
-        E: serde::de::Error,
-    {
-        self.visit_str(value.as_ref())
-    }
-}
 
 #[derive(serde::Serialize, serde::Deserialize, Debug, PartialEq, Eq, Clone, Default)]
 pub struct ChDbConfig {
