+++ evm_loader/program/src/evm/tracing/event_listener/listener_vm_tracer.rs
@@ -1,9 +1,6 @@
+use crate::evm::tracing::event_listener::trace::{MemoryDiff, StorageDiff, VMTracer};
+use crate::evm::{Context, ExitStatus};
 use ethnum::U256;
-
-use evm_loader::evm::{Context, ExitStatus};
-
-use crate::event_listener::trace::VMTracer;
-use crate::event_listener::trace::{MemoryDiff, StorageDiff};
 
 use super::vm_tracer::VmTracer;
