+++ evm_loader/program/src/evm/tracing/event_listener/mod.rs
@@ -5,7 +5,7 @@ pub mod trace;
 pub mod tracer;
 mod vm_tracer;
 
-use evm_loader::evm::tracing::{Event, EventListener};
+use crate::evm::tracing::{Event, EventListener};
 use {listener_tracer::ListenerTracer, listener_vm_tracer::ListenerVmTracer, tracer::Tracer};
 
 impl EventListener for Tracer {
