+++ evm_loader/program/src/evm/tracing/event_listener/vm_tracer.rs
@@ -1,4 +1,4 @@
-use crate::event_listener::trace::{ExecutiveVMTracer, MemoryDiff, StorageDiff};
+use crate::evm::tracing::event_listener::trace::{ExecutiveVMTracer, MemoryDiff, StorageDiff};
 use ethnum::U256;
 
 #[derive(Debug, Default, Clone)]
